chore: move onSubmit to BaseModule
